requirements.txt
Flask==2.0.2
requests>=2.23.0
python-dotenv==0.17.1
